--- ribbit-front/src/container/Regiseter/Login.jsx
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-
-class Login extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
-  render() {
-    return <div />;
-  }
-}
-
-export default Login;
